update for ex_day 4

+++ src/main/java/com/example/yootday1/service/impl/CourseServiceImpl.java
@@ -1,9 +1,13 @@
 package com.example.yootday1.service.impl;
 
+import com.example.yootday1.common.exception.NotFoundException;
 import com.example.yootday1.domain.entity.Course;
+import com.example.yootday1.dto.course.CourseResponse;
+import com.example.yootday1.dto.course.CourseUpsertRequest;
 import com.example.yootday1.repository.CourseRepository;
 import com.example.yootday1.service.CourseService;
 import lombok.RequiredArgsConstructor;
+import org.modelmapper.ModelMapper;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
@@ -13,25 +17,44 @@ import java.util.Optional;
 @RequiredArgsConstructor
 public class CourseServiceImpl implements CourseService {
     private final CourseRepository courseRepository;
+    private final ModelMapper mapper;
 
-    public List<Course> findAll(){
-        return courseRepository.findAll();
+    private CourseResponse map(Course course){
+        return mapper.map(course, CourseResponse.class);
     }
 
-    public Optional<Course> findById(Long id){
-        return courseRepository.findById(id);
+    public List<CourseResponse> findAll(){
+        return courseRepository.findAll().stream()
+                .map(r->map(r)).toList();
     }
 
-    public Course save(Course course){
-        return courseRepository.save(course);
+    public List<CourseResponse> findByCourseActive(){
+        return courseRepository.findByCourseActive().stream()
+                .map(r->map(r)).toList();
     }
 
-    public Course updateCourse(Long id, Course course){
+    public Optional<CourseResponse> findById(Long id){
+        return courseRepository.findById(id).map(this::map);
+    }
+
+    public CourseResponse create(CourseUpsertRequest req){
+        Course course = mapper.map(req, Course.class);
+        Course response = courseRepository.save(course);
+        return map(response);
+    }
+
+    public CourseResponse update(Long id, CourseUpsertRequest req){
+        Course course = mapper.map(req, Course.class);
         course.setId(id);
-        return courseRepository.save(course);
+        Course response = courseRepository.save(course);
+        return map(response);
     }
 
-    public void deleteCourse(Long id){
-        courseRepository.deleteById(id);
+    public void delete(Long id) throws NotFoundException {
+        if (courseRepository.existsById(id)){
+            courseRepository.deleteById(id);
+        }else {
+            throw new NotFoundException("Delete error");
+        }
     }
 }
